+++ R/req-perform.R
@@ -32,6 +32,9 @@
 #'   * 1: show headers
 #'   * 2: show headers and bodies
 #'   * 3: show headers, bodies, and curl status messages.
+#'
+#'   Use [with_verbosity()] to control the verbosity of requests that
+#'   you can't affect directly.
 #' @returns If request is successful (i.e. the request was successfully
 #'   performed and a response with HTTP status code <400 was recieved), an HTTP
 #'   [response]; otherwise throws an error. Override this behaviour with
@@ -43,10 +46,11 @@
 req_perform <- function(
       req,
       path = NULL,
-      verbosity = getOption("httr2_verbosity", 0L),
+      verbosity = NULL,
       mock = getOption("httr2_mock", NULL)
   ) {
   check_request(req)
+  verbosity <- verbosity %||% httr2_verbosity()
 
   if (!is.null(mock)) {
     mock <- as_function(mock)
@@ -138,7 +142,7 @@ req_perform1 <- function(req, path = NULL, handle = NULL) {
 
 req_verbosity <- function(req, verbosity) {
   if (!is_integerish(verbosity, n = 1) || verbosity < 0 || verbosity > 3) {
-    abort("verbosity must 0, 1, 2, or 3")
+    abort("`verbosity` must 0, 1, 2, or 3")
   }
 
   switch(verbosity + 1,
@@ -290,3 +294,9 @@ req_handle <- function(req) {
 
 new_path <- function(x) structure(x, class = "httr2_path")
 is_path <- function(x) inherits(x, "httr2_path")
+
+check_verbosity <- function(x) {
+  if (!is_integerish(x, n = 1)) {
+    abort("`verbosity` must be a single integer")
+  }
+}